Fix build on Windows
There was a /Machine:x64 as compiler flag which was making the build
failed when build for x86 (the default environment on Visual Studio).
By removing this flag the link should work for both x86 and x64.

I have also add an output name for the .pdb associate with one .obj on
the compiler command line. Before all obj file were sharing the same
the same .pdb file and it was creating some issues during the generation
of the final .pdb.